requirements.txt
aiofiles==0.8.0
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
ci-info==0.2.0
click==8.0.3
colorama==0.4.4
configobj==5.0.6
configparser==5.1.0
etelemetry==0.2.2
filelock==3.4.0
fitz==0.0.1.dev2
frontend==0.0.3
future==0.18.2
h11==0.12.0
httplib2==0.20.2
idna==3.3
isodate==0.6.0
itsdangerous==2.0.1
lxml==4.6.4
networkx==2.6.3
nibabel==3.2.1
nipype==1.7.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pathlib==1.0.1
pdf2image==1.16.0
Pillow==8.4.0
prov==2.0.0
pydot==1.4.2
pyparsing==3.0.6
pytesseract==0.3.8
python-dateutil==2.8.2
pytils==0.3
pytz==2021.3
pyxnat==1.4
rdflib==6.0.2
requests==2.26.0
scipy==1.7.3
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tools==0.1.9
traits==6.3.2
urllib3==1.26.7
uvicorn==0.15.0
